requirements.txt
Flask==2.2.5
googletrans==4.0.0-rc1
torch==2.6.0

# LangChain
# LangChain
langchain==0.1.16
langchain-community>=0.0.30
langchain-experimental>=0.0.20
langchain-google-genai>=0.0.7
google-generativeai>=0.4.0

# Pydantic compatible
pydantic>=2.0

# Data
pandas

# NLP & Embeddings
sentence-transformers==2.6.1
transformers==4.37.0
huggingface-hub==0.20.3


